A variety of additions
I added the header tag to the home section, I added the sections for each payment model with two games under each with an image and an imbedded video in a details tag for each game. I also added the Recommendation section that featured a form that used radio attribute input in addition to a text box for entering an email address. Finally, I added a footer with a made-up web address.